+++ scripts/sandbox/echo-cause-container-restart.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Copyright 2024 Expedia Group, Inc.
+# Copyright 2025 Expedia Group, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.